Protoype v.01

+++ sunrin-pong-prj/cVector2.h
@@ -2,6 +2,6 @@
 
 class cVector2 {
 public:
-	int x, y;
-	cVector2(int x, int y) { this->x = x; this->y = y; }
+	float x, y;
+	cVector2(float x, float y) { this->x = x; this->y = y; }
 };
